feat: 모임상에 SectionModel 생성
